Fix some bugs with remembered set scanning
The problem is that we cannot scan an entire region.  We have to scan
only the portion of the region that has already been allocated.  This
commit is a work in progress.  There is still at least one place where
it appears I am scanning too much of the old-gen region.  Also, this
commit has some printf statements to help diagnose the whereabouts of
the problem.

I'll continue working on this tomorrow.