Replace lame remembered set scanning
This patch implements the originally planned remembered set scanning
which had been set aside months ago because it was perceived to be too
complex, and there was a risk that implementing it now would result in
excessive delays in milestone demonstrations.  We have subsequently
discovered that the "quick and dirty" remembered set scanning that was
implemented as a replacements runs so slowly that it defeats all of the
intended benefits of generational shenandoah.  This new implementation
appears to run at least 100 times faster.